feat(interpreter): getelementptr added. asm builder developing.

+++ src/switch.rs
@@ -3,30 +3,30 @@ use crate::interpreter::{interpret_func, InterpreterContext};
 use crate::mem::{FUNC};
 
 #[unsafe(no_mangle)]
-pub extern "C" fn asm_call_fn_handler(fn_index: usize, ctx: *mut usize) {
+pub extern "C" fn asm_call_fn_handler(fn_index: u64, ctx: *mut u64) {
     // todo
-    let func = &FUNC.exclusive_access()[fn_index];
+    let func = &FUNC.exclusive_access()[fn_index as usize];
     let mut int_ctx = InterpreterContext::new();
     let ret = interpret_func(func.clone(), &mut int_ctx);
+    
     unsafe extern "C" {
-        fn __interpreter_ret_asm(ret: i64, ctx: *mut usize);
+        fn __interpreter_ret_asm(ret: i64, ctx: *mut u64);
     }
     unsafe {
-
         __interpreter_ret_asm(ret, ctx);
     }
     panic!("should not reach here");
 }
 
-pub fn interpreter_call_asm(fn_index: usize, para_num: usize, para_ptr: usize) -> isize {
+pub fn interpreter_call_asm(fn_index: u64, para_num: u64, para_ptr: u64) -> i64 {
     let new_sp = alloc_stack();
     unsafe extern "C" {
         fn __interpreter_call_asm(
-            fn_ptr: usize,
-            para_num: usize,
-            para_ptr: usize,
-            new_sp: usize,
-        ) -> isize;
+            fn_ptr: u64,
+            para_num: u64,
+            para_ptr: u64,
+            new_sp: u64,
+        ) -> i64;
     }
     let ret = unsafe {
         __interpreter_call_asm(fn_index, para_num, para_ptr, new_sp)
